fixed 2 bugs 'Cast one of the operands of this subtraction operation to a long'.

+++ src/main/java/org/apache/commons/text/StrBuilder.java
@@ -153,7 +153,7 @@ public class StrBuilder implements CharSequence, Appendable, Serializable, Build
         /** {@inheritDoc} */
         @Override
         public long skip(long n) {
-            if (pos + n > StrBuilder.this.size()) {
+            if (pos + n > (long) StrBuilder.this.size()) {
                 n = StrBuilder.this.size() - pos;
             }
             if (n < 0) {
